--- config/appeal_status/mock_get_appeals_responses.yml
@@ -1,5 +1,0 @@
-users:
-  "796126859":
-    appeals:
-      - appeal_status: SOC_IN_PROGRESS
-      - appeal_status: FORM_9_NEEDED
