+++ lib/appeals_status/models/appeals.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+require 'appeals_status/models/appeal'
+
+module AppealStatus
+  module Models
+    class Appeals
+      include Virtus.model
+      attribute :appeals, Array[AppealStatus::Models::Appeal]
+    end
+  end
+end
