+++ lib/appeals_status/responses/get_appeals_response.rb
