+++ spec/lib/appeals_status/responses/get_appeals_response_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 require 'rails_helper'
-require 'appeal_status/responses/get_appeals_response'
+require 'appeals_status/responses/get_appeals_response'
 
 describe AppealStatus::Responses::GetAppealsResponse do
 end
